# api/gen/server/quotes/model_item.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Quotes API
 *
 * API for quotes
 *
 * API version: 1.0.0
 */

package quotes

// Item - Quote item
type Item struct {
	Id string `json:"id"`

	Segment string `json:"segment"`

	Price float32 `json:"price,omitempty"`

	Tax float32 `json:"tax,omitempty"`
}

// AssertItemRequired checks if the required fields are not zero-ed
func AssertItemRequired(obj Item) error {
	elements := map[string]interface{}{
		"id":      obj.Id,
		"segment": obj.Segment,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertItemConstraints checks if the values respects the defined constraints
func AssertItemConstraints(obj Item) error {
	return nil
}