# api/gen/server/quotes/model_quote.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Quotes API
 *
 * API for quotes
 *
 * API version: 1.0.0
 */

package quotes

// Quote - Quote
type Quote struct {

	// Quote ID
	Id int64 `json:"id,omitempty"`

	// Customer ID
	CustomerId int64 `json:"customer_id"`

	// Quote description
	Description string `json:"description"`

	// Quote status
	Status string `json:"status,omitempty"`

	Items []Item `json:"items,omitempty"`
}

// AssertQuoteRequired checks if the required fields are not zero-ed
func AssertQuoteRequired(obj Quote) error {
	elements := map[string]interface{}{
		"customer_id": obj.CustomerId,
		"description": obj.Description,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	for _, el := range obj.Items {
		if err := AssertItemRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertQuoteConstraints checks if the values respects the defined constraints
func AssertQuoteConstraints(obj Quote) error {
	for _, el := range obj.Items {
		if err := AssertItemConstraints(el); err != nil {
			return err
		}
	}
	return nil
}